新增公告管理

+++ yyms-web/src/main/java/net/laoyeye/yyms/controller/BaseController.java
@@ -30,7 +30,7 @@ public class BaseController{
      */
     protected static Result result(Page jpaPage) {
 
-        return new Result(jpaPage.getContent(), (long)jpaPage.getTotalElements());
+        return new Result(jpaPage.getContent(), jpaPage.getTotalElements());
     }
 
 }
